fix: retry with git checkout
Works around rate limiting on git. We may want to up this count but it is not a bad idea to still have retries. Turns out this hits the network. Not that surprising in retrospect (git checkout can sync to find branches on the remote), but I usually think of it in the fully local DB case so didn't think of adding this before