chore: Remove "as unknown" casts for ABIs where possible
Adds missing path property to struct types and removes "as unknown"
casts for going from json files to ABI types. Note that the cast is
still needed since ts loads JSON files with values understood as
strings, so without the cast the type checker complains about
`visibility` being a string rather than an `ABIVisibility`. And changing
the visibility type from an enum to a union type doesn't cut it, see
[this ts issue](https://github.com/microsoft/TypeScript/issues/26552)
for more info.