BROKEN BROKEN NOIR BUG:
Running `yarn noir:build && yarn test` on the noir-private-kernel-lib on this commit will give you a failure with the message `    Cannot satisfy constraint`.

The problem is `function_tree_root_from_siblings` which calls `root_from_sibling_path` which fails on L32, the indexing of `indices[i]`.

It is not an index out of bounds error, and seems to only happen with this test