feat: Bring back bb.js sync api. (#3345)
* We once again codegen a sync version of bb.js api, and use it instead
of the hand rolled Pedersen class.
* This improves upon how we called into the wasm. Previously when it was
in a worker, every mem copy/slice and call was an individual call over
the boundary. Now the binder is directly in the wasm wrapper, and there
is a separation of "enriched type" construction. i.e. the call over the
boundary only contains primitives, and contains all the necessary info,
and the returned value contains all the sliced memory, then the
codegened api does the enriching.
* The above also means we don't need two different binders/heap
allocator handlers for async/sync apis.
* It should also mean we never worry about two async calls being made
over the boundary at once and corrupting the shared scratch space.
* We construct a singleton of the sync api on demand in bb.js, so the
instance can be used repeatedly.
* At present it's just used in `foundation/crypto/pedersn`, but can now
start leveraging it in Aes/Ecdsa/Grumpkin classes etc.