chore: Less circuits wasm, more bb.js wasm. (#3347)
* Branch name perhaps not totally honest, as we're not using the codegen
api. A simpler stepping stone was to substitute in the bb.js wasm for
the circuits wasm in the various classes that used it.
* Switching to use api for making the wasm calls will require updating
the bindings to follow the bindgen calling convention. I've done this
for aes, but can do the others later.
* No more async needed to create aes/schnorr/grumpkin etc.
* We can probably get rid of the classes themselves soon and just have
free functions.

If we seach the codebase now for `await CircuitsWasm.get()` there's only
a handful of lines left, mostly related to what I think are deprecated
circuit simulation calls. Once we get rid of these, we can get rid of
building the circuits in CI.