chore: Bring boxes back to CI. Build and run using docker/docker-compose. (#3727)
Main gist is as title. Details below.
* Rename `noir-acir-tests` -> `noir-compile-acir-tests` to make it more
clear this is compile job, not a test job.
* `boxes` files are injected into `yarn-project-base` much like other
artefacts, as they're need for cli unboxing. Job is called
`boxes-files`.
* Added myself as a CODEOWNER for `build-system` and `build_manifest`
changes (just means I'm auto-added as reviewer).
* `PATH` adjustment in `aztec-nargo` to ensure it can find rootless
docker.
* `boxes` is cleaned up a lot. Removed a lot of "development time"
specifics. Made it into a workspace so we can build all the boxes as a
single job, and test them individually as part of a docker-compose run.
* The root `package.json` contains the resolutions to development
dependencies. As this isn't a part of the unboxed box, no adjustment
needs to be made r.e. resolutions.
* We still update the `@aztec/*` package versions as we unbox, but in
theory this could now be done by release please.
* However, there is still the issue of `Nargo.toml` which needs a more
significant change to point to github dependencies. If we could devise a
more cunning mechanism here, it's possible we could avoid needing to
adjust the projects at unbox time at all.
* HOWEVER, there is another case which I've supported here, *perhaps*
overkill, which is running unbox at development time. Such an unboxing
needs to carry the dependencies into the box and make several tweaks.
This has the nice benefit of being able to provide the full
functionality without having to have published anything, but adds a
couple of dozen lines of code otherwise not needed.
* Gets rid of scripts like `compile:local` which make no sense when
unboxed. Instead allow overriding `aztec-cli` via env var.
* Remove project references from `tsconfig.json` because we simply
assume to have built `yarn-project` in the development case.
* Remove committed noir artefacts and json, and instead build them at CI
time, or as part of `yarn build` once unboxed.
* Updated rotted noir code. TODO: replace with symlink to `src` so no
need to keep in sync? All symlinks are derefed at unbox time.
* Provide a simple `docker-compose` file for running any box against the
sandbox.
* Get rid of `tsconfig.dest.json` stuff. Simplify some webpack stuff
that was just copy pasted. (more needs to be done to strip this back to
be minimal, get rid of polyfills etc).
* Minor build system tweaks to fix dependencies when running
docker-compose.