feat: Goblinize the final ecc ops in ZM (#3741)
In the initial implementation of ZM I forgot to goblinize a couple of
the final group operations in Zeromorph which was causing the gate count
of the recursive verifier to blow up. We fix this. We implement the
addition, subtraction, and single scalar multiplication operation
inefficiently using the batch_mul methods. We disable some member
functions that do not use Goblin when they could, and leave references
to https://github.com/AztecProtocol/barretenberg/issues/707 as a future
optimization task.

---------

Co-authored-by: codygunton <codygunton@gmail.com>