fix: Flaky deployment test
The test 'should not deploy a contract which failed the public part of the execution' depended on having a good and a bad tx. But the bad tx accidentally dependend on the good one, since it required the good one to register the class. If the txs were accidentally run out of order, the bad tx would fail earlier than we needed. So this fix ensures there's no dependency between the two.