chore: Use inotifywait to run generate in yarn-project
Replaces `yarn build:dev` at the root of yarn-packages with a script
that leverages inotifywait to re-run `yarn generate` in the packages
where it's needed. To achieve this, we define specialized `generate`
commands based on the dependencies (l1 artifacts, noir contracts, or
noir protocol circuits) and run the one depending on the modified
folder.

Since typescript watch doesn't seem to like if we modify too many files
at a time, we kill it while we're doing codegen, and then restert it.
Also, to make sure we don't run codegen before the compilation of the
dependent artifacts is completed, we debounce the call by a few seconds.