feat: enable honk_recursion through acir (#6719)
We want to be able to verify honk proofs without adding a new opcode.
The workaround that this PR introduces is adding a flag that determines
whether to create a plonk vs a honk recursion constraint based on which
proof system we're using to prove. If we are using ultra honk, we will
generate a honk recursion constraint and the same for plonk.

This will need to be reverted after the offsite:
https://github.com/AztecProtocol/barretenberg/issues/1013.

Please read [contributing guidelines](CONTRIBUTING.md) and remove this
line.