feat: Poor man's CLI block explorer
Adds a command line block explorer in the CLI via a new `get-block`
command, with an optional `--follow` for streaming new blocks. Required
adding a new call to the PXE to return a contract artifact (as opposed
to a class) and a new note filter to retrieve notes by nullifier.