feat: Do not evict tx objects from p2p tx pool immediately
Tweaks the p2p client so it waits a configurable number of blocks before
deleting proven txs from its pool. This can help with reorgs or
troubleshooting, and it also allows slow provers to get the tx objects
they need in a scenario where multiple provers can submit a proof for
the same block.