fix: Anvil block timestamp
Anvil uses the time of the forked block as its timestamp as its
baseline. Subsequent blocks have a timestamp equal to that baseline plus
the time that passed since the fork was created.

To fix it, we set the next block timestamp manually to the current date
and immediately trigger a block mine.