feat: Sort proving jobs by epoch number
The proving jobs queue sorted jobs in FIFO order. When we added support
for proving multiple blocks simultaneously, this led to an inefficiency:
if jobs for two blocks were pushed into the queue, the merge and root
jobs for the first block would not be computed until the base jobs for
the second one were done.

This PR changes the queue so it works by priority, where priority is
based on an epoch number, defined as the block number for simplicity.

This also renames the "fifo" export in foundation to "queue", since we
now support priority in addition to fifo queues.