chore!: Sequencer no longer proves
Removes the code path where the sequencer would submit proofs for a
given block, so this reponsibility is now exclusive of the prover-node.
This allows us to remove the prover instance from the aztec node
altogether, and drop the `PROVER_DISABLED` and `SEQ_SKIP_SUBMIT_PROOFS`
env vars since are no longer needed.