feat: Max pending jobs in prover node
Adds a max number of pending block-proving jobs to the prover node. When
reached, the prover node stops proving, and resumes once a job finishes.
Also adds error handling to the creation of a new proving job, so if a
block-proving-job fails to start, it is skipped instead of blocking the
prover node.