chore: Revert to current date as timestamp for devnet
When running devnet, using the slot timestamps instead of current
timestmap meant that blocks were often generated with a timestamp much
into the future. This screwed up settings such as the min and max time
between blocks for a sequencer configuration. In addition, the custom
block timestamp interval in anvil meant that we could no longer rely on
L1 timestamps for measuring time, which was needed for tracking the time
between a block being submitted and its proof.

This PR changes timestamps if devnet is set so that we use the current
timestamp for building blocks, do not alter anvil's block timestamp, and
drop time validations.