feat: Block cheat codes in anvil
Adds a second api to the mainnet-fork, located at
`public-PUBLIC_API_KEY`, in which all cheat codes are blocked. We define
a cheat code as a method in the evm, hardhat, or anvil namespaces. We
use njs for parsing the JSON RPC request body and testing the method
namespace.

The PUBLIC_API_KEY is only set for provernet, for other networks, we
load the same API_KEY secret as always.

This PR also enables logging for anvil, including a logrotate config to
ensure we don't fill up the disk just with logs.