fix: Split stores per component and split merkle tree operations (#8299)
This PR packs two changesets:

1. Spins off a `MerkleTreeAdminOperations` from `MerkleTreeOperations`,
to have an interface that does not expose methods that commit changes to
the underlying store. World state now exposes a method to get an
`ephemeralFork` without these operations, that can be used for answering
public simulation requests. Note that we do not yet enforce that no
changes go into the store, that requires more changes to have "readonly"
versions of the trees and store all the way down.
2. Moves creation of the underlying data store to the factory of each
component, so each creates a new db as needed, instead of sharing a
single one. This allows to keep separate db files for p2p, archive, and
world state. As a bonus, it makes forking world state cheaper.