chore(ci): skip vk generation on `protocol-circuits-gates-report` and `noir-format` (#8398)
We're currently generating the verification keys for the protocol
circuits in order to run the formatter on them and count the number of
gates in the circuit. This is then much slower than it needs to be when
we change the noir compiler/barretenberg such that the vk changes.

This PR moves these targets to pull from the `+source` rather than
`+build-protocol-circuits` target so we can avoid this.