chore: Moves add gate out of aux (#8541)
Resolves https://github.com/AztecProtocol/barretenberg/issues/913.

Moves the add gate in aux to the arithmetic block instead. Adds some
tests to check the partitioning of selectors into their blocks.

This used to be blocked by the recursion limit in Plonk (as in this change would've bumped us over the limit), but we don't
care about that anymore (because it's double_verify_proof already goes over 2^19).