refactor: reduce redundant event fetching (#8628)
Fixes #8565. 

- Adds a `status` function to the rollup contract
- Archiver will update `lastSynchedL1Block` for blocks and skip looking
for events if the rollup contract state tells it that there will be no
events.

I was running into some oddities with the timing of updates to the
`lastSynchedL1Block` for the blockstore. Namely, seems like it was only
updated if events were found, leading to a large number of "re-fetch" as
we keep looking over the same ranges until an event swing by.