chore(ci): scriptify local earthfile
Longer term, this file doesn't make a ton of sense, only really being
used like a Makefile that can give its valid targets. Going forward we
should find another way to get the list of targets, but for now just
refactoring the low-handing fruit so these are functions that can be
called by CI jobs directly. That will happen in a future pass.