chore: migrate higher-level APIs for barretenberg to bb.js (#8677)
This PR moves a bunch of the code in `noir_js_backend_barretenberg` into
`bb.js` . I've kept this as separate classes for now but we can likely
migate more functionality over to the standard `Barretenberg` class
(especially with UP being deprecated as UH has less external state which
needs to be managed.)

`noir_js_backend_barretenberg` still exists as we still need to split
off public inputs, etc. to avoid this being a breaking change or
creating an inconsistency in `bb.js`. We can look at how to address this
in future.