refactor: archiver identifies prune (#8666)
Fixes #8620, by using the status function to figure out when a prune
have happened and unwind state. It do NOT tell the world state
synchronizer about the unwinding and that should be addressed as part of
#8665.

Adds tests to `e2e_synching` making sure that an archiver will correctly
catch the prune happening and delete the blocks and accompanying data.
Also adds e2e_synching to earthfile.

Removes the `proven_store` since we learn this from the status and it is
not more closely linked to the fetching of blocks.

Note that we are not handling L1 re-orgs specifically in here, but they
will for some cases be caught as well. But not all, so there is still
#8621 as well.

Fixes an issue where asking for block 0 would return block 1, as there
is no block 0. Explicitly return undefined if block 0 is requested now,
and gives an error if using `getBlocks` from a block that is before the
initial block.