fix: Do not start block root rollup proof before block is built (#8952)
We were running into a race condition where, under simulated circuits,
the block root rollup was being proven _before_ the block was actually
built. This triggered an error when trying to access the built block for
validating the output of the root rollup circuit:


https://github.com/AztecProtocol/aztec-packages/blob/7ef1643218356d22d09601269f346927694e22d7/yarn-project/prover-client/src/orchestrator/orchestrator.ts#L919

See [this
run](https://github.com/AztecProtocol/aztec-packages/actions/runs/11112244458/job/30876295492?pr=8776#step:5:802)
for an example.