chore: Spartan kubernetes documentation (#9012)
This PR adds a `README` file with more information about the Spartan
Kubernetes cluster.

The PR also includes an update to the CI job which deploys networks when
code is merged into `staging` and `production` branches. This update is
needed because the original `spartan_cluster` was redeployed and is now
simply `spartan`.