fix: Archiver getBlocksForEpoch and EpochProvingJob on block 1
Fixes a few issues around proving:

Archiver was returning block headers in reverse order when filtering
by epoch. This hadn't popped up in e2e tests before because we were
testing with single-block epochs. We now changed the e2e_prover test to
have two blocks in the epoch.

Epoch proving job would fail to prove an epoch that started in block one
because it requested the header of the previous block to begin, which
was undefined. That edge case is now handled.

Bond manager would trigger a top up when the amount to top up was
exactly zero.