fix: public write batches (#8986)
This PR restores the batched approach to synching public data writes.

One thing that's different to the previous approach is that now the
inserts into the public data tree do not block inserts into the other
trees. This was achieved using a recursive approach where each batch
enqueues the next batch to be inserted.