feat: Connect the public inputs but not the proof in ivc recursion constraints (#8973)
The inputs to an ACIR ivc recursion constraint include the VK, the
public inputs, and the proof (without public inputs). The VK and public
inputs are expected to be used in aztec-y logic directly in the noir
protocol programs, but the proof is not. Previously, we connected the
witnesses for all three of these components to the counterpart used in
the backend recursive verifier. This requires knowledge of the proof
size, could be misleading and is unnecessary. With this PR, we only
create and connect witnesses for the VK and public inputs and not the
proof when constructing an ivc recursion constraint.