fix(avm): guard against inf (#9081)
This is a bit of a hack for now. The bb wasm interface represents
infinity in different ways - none of which utilise the boolean
`isInfinite` flag. Since aztec-nr shouldn't run into this issue, we
solve it in an avm-specific way.

Specifically, we just guard the opcodes to treat infinities separately -
without calling the bb wasm interface

Closes #9020