feat: persistent storage edit for anvil node (#9089)
This PR includes the following changes: 

# Migrated Aztec Grafana Dashboard

After the Grafana stack was split out into a separate helm chart ( #9037
), a new Grafana instance has been installed to the Spartan kubernetes
cluster. This installation is inside a new `metrics` namespace and is
not tied to any single deployment.

To allow using the existing dashboard from grafana.aztec.network, the
`uid` of the datasource used for each of the dashboard panels has been
updated to use the new Grafana instance name.

`"uid": "aztec-node-metrics"` >> `"uid": "spartan-metrics-prometheus"`

The dashboard has since been imported to the new grafana instance
(available at spartan.grafana.aztec.network). The DNS name is expected
to change in the future, once production networks run from kubernetes
and the current grafana.aztec.network instance is retired.

# Persistent Grafana Stack Storage

This PR adds persistent storage for the portions of the Grafana stack
that don't include it by default (specifically the Grafana dashboard
itself).

Additionally, a load balancer configuration is added to expose Grafana
publicly, as well as additional annotations that allow it to support
https for the CNAME spartan.grafana.aztec.network using an AWS issued
certificate (pending final AWS approval).

# Anvil Persistent Storage

By default, Anvil stores its activity in memory
([ref](https://book.getfoundry.sh/reference/cli/anvil)). This PR mounts
a persistent volume and uses the `--load-state` and `--dump-state` flags
to export the current network state at fixed intervals.

In testing, Anvil does not allow mounting a persistent volume to either
the `/` or `/home/foundry` directories. This method compensates for
that. It should be noted that an L1 gap is still possible if Anvil dies
before the most recent state is exported.

#