feat: modify private calldata to use public keys (#9276)
In this PR we are modifying private calldata to use public keys instead
of only the public keys hash. This is in preparation for the new address
changes in which we private calldata is needed to validate a contract
address in the kernels in `validate_contract_address`
(validate_contract_address uses information from the private calldata
struct), and we need to constrain that a contract address can be
recomputed using its entire set of public keys with
`AztecAddress::compute_from_private_function`.