feat: introduce default public keys and replace empty public keys (#9277)
This PR further prepares our new address changes. Because we need to
prove correct derivation of the address, and that the definition of the
address has changed to be h (the pre-address) * G added to Ivpk, we need
this Ivpk to be an actual point on the curve because if it isn't, any
attempted derivation of the address will fail. Because of this we are
now removing the concept of "zeroed" or "empty" keys because they are no
longer valid in an address, as they would never be able to be used.

We replace our empty public keys in favor of the concept of default
public keys, and we replace all of them instead of only the ivpk for
consistency. These "default" keys are keys that derived from
 
"az_null_npk"
"az_null_ivpk"
"az_null_ovpk"
"az_null_tpk"
as bytes, hashed to curve using
`grumpkin::g1::affine_element::hash_to_curve(<X>, 0)`;