fix(k8s): boot node long sync (#9610)
If the boot node needs to catch up after a big sync, it can be killed by
the liveness probe before it has a chance to finish

Hit a major somewhat unexplained gotcha - for some reason, when
introducing a startupProbe, self-introspecting DNS did NOT work. I
couldn't find documentation for it, but it seems that DNS waits for
startupProbe. I agreed with Mitch that this works just fine using
POD_IP.