chore: Validate RPC inputs (#9672)
**Adds schemas for every API.** Every API exposed via JSON RPC now
requires a zod schema (see #9656 for more context on the rationale for
this change). All schemas are in `circuit-types/interfaces`, and look
like:


https://github.com/AztecProtocol/aztec-packages/blob/3e78ec721285fcd533cff61329a8e156958e2d65/yarn-project/circuit-types/src/interfaces/prover-node.ts#L33-L45

These schemas are type-checked against the interface via the
`ApiSchemaFor` utility type, so if the interface changes, schemas are
required by the compiler to change as well. Schemas are now used in the
JSON RPC server to 1) identify which methods are exposed (so we no
longer need the method disallowlist) and 2) parse their arguments. The
JSON RPC server, once it has identified the method to be called, grabs
the arguments schema and funnels the result of a vanilla JSON parse
through it.

Every type or struct that is exposed via an interface now has an
associated schema, which is referenced in the API for parsing. Schemas
both validate input and hydrate instances. This means that we no longer
set a `type` property to identify how to hydrate each object in a
request during deserialization, which was a security risk.


https://github.com/AztecProtocol/aztec-packages/blob/3e78ec721285fcd533cff61329a8e156958e2d65/yarn-project/circuit-types/src/l2_block.ts#L24-L32

Schemas are also used in the JSON RPC client for deserializing the
result types. Again, this lets us remove the `type` parameter from all
serialized entities, though this is still present in since it is
required by the `TypeRegistry` (still to be removed) which is only used
in the snapshot manager.

All schemas are tested via mini integration tests. These tests define a
mock implementation for each service, use it for setting up a JSON RPC
server, starting it in a free port, and test calling every method
through JSON RPC.


https://github.com/AztecProtocol/aztec-packages/blob/3e78ec721285fcd533cff61329a8e156958e2d65/yarn-project/circuit-types/src/interfaces/prover-node.test.ts#L12-L31

These changes prompted other changes. For instance, we introduced the
following changes to APIs:

- `ProvingJobSource.rejectProvingJob` now accepts a reason `string`
instead of an `Error` type
- `PXE.getEvents(type)` is removed in favor of `PXE.getEncryptedEvents`
and `PXE.getUnencryptedEvents` since both methods required different
arguments

We also removed service-management methods (ie `stop`) from interfaces.
We were inadvertently calling `stop` on remote instances over http when
we shouldn't have. We also typed some previously untyped interfaces,
such as the TXE's.

Fixes #9455