fix: add zod parsing for generated contract artifacts (#9905)
This PR enables us to enforce deterministic contract artifact JSON
stringifying. This is because now when we generate our contract
artifact, we pass it through our zod parser, which enforces ordering in
the object, and enforces correct ordering of props in objects. This is
related to #6021, which we need for deterministic artifact hashes.