fix: Allow unwinding multiple empty blocks
The archiver kv store used the block **body** hash as identifier to
store the block bodies. This means that two empty blocks would have the
same identifier. So, when unwinding (ie deleting) two or more empty
blocks, the first one would remove the body that corresponds to an empty
block, and the second one would fail with "Body could not be retrieved".

This fixes the issue by using the block hash, guaranteed to be unique,
as a key to store the block bodies.