feat: Single commitment key allocation in CIVC (#9974)
Previously we allocated the BN254 commitment key freely (I counted 11
times in one small ClientIVC test). This is unnecessary and could lead
to memory fragmentation. This PR implements size functions on a
`TraceSetting` object and, when structured traces are used, allocates
the commitment key at CIVC construction time. It passes this along to
dependent classes via a shared pointer.

I didn't handle the case of Grumpkin since it's not an issue.

I was curious to validate the effect in the browser directly through the
browser app in the ivc-integration test suite. Though it's tangential, I
updated that app to display console logs on the page for easy sharing.