feat: swap polys to facilitate dynamic trace overflow (#9976)
Update PG/ClientIvc so that during accumulation we can handle a circuit
that overflows the nominal structured trace (and potentially increases
the dyadic size of the accumulator) without knowing about the size of
the overflow in advance. (A previous PR makes it possible to overflow
arbitrarily as long as ClientIvc is initialized with the proper overflow
capacity). Ensure the ExecutionTraceTracker is updated correctly when
encountering a circuit that overflows and also ensure that the tracker
is used appropriately in Protogalaxy.

---------

Co-authored-by: maramihali <mara@aztecprotocol.com>