feat: Process blocks in parallel during epoch proving
Instead of processing blocks in sequencer during epoch proving and
trigger proving jobs as each block is processed, we fetch the state
prior to each block from a world state following the pending (not
proven) chain and process each block (ie execute public functions) in
parallel. This means tx execution is less of a bottleneck for proving.