chore: Parallelise construction of perturbator coefficients at each level (#10304)
When constructing the perturbator via the tree technique we can
parallelise the construction of the coefficients at each level which is
one of the culprits of performance degrading when using an ambient trace
of size 2^20. We see a ~1s improvement in performance doing this.

2^20 trace (`EXAMPLE_20`) in master for `ClientIVCBench/Full/6` was
`38114 ms` and is now `37310 ms`. For the defacto
`CLIENT_IVC_BENCH_STRUCTURE` which gives 2^19 finalised circuits we go
from `29496ms` to `29188 ms` so not as impactful but also the
performance doesn't degrade.

I have also benchmarked the actual computation of coefficients and it is
neglegible regardless of the ambient trace size (2^20 vs 2^19) .