feat: release please for release branch (#10467)
Allows us to have release please PRs for our `release/*` branches. 

When we merge one of the release please PRs on a `release/*` branch
(e.g. `release/sassy-salamander`), the following happens:

- We create new images in dockerhub at e.g.
`aztecprotocol/aztec:sassy-salamander`
- We bump the `devnet` images in dockerhub at e.g.
`aztecprotocol/aztec:devnet`
- We create a regular (semantic versioned) npm release but tagged
`devnet` (as opposed to `latest`)
- We publish aztec-up to version `sassy-salamander`

All other existing publishing flows are unchanged.

Note: this implies that *before* merging the release PR, we have
deployed a new devnet, and the public `sassy-salamander` network.

Note: The
[issue](https://github.com/AztecProtocol/aztec-packages/issues/10160)
states that if there are no changes between the `sassy-salamander`
branch and `master`, it will simply move tags forward. This does *not*
accomplish that. If that is really desired we should create a separate
ticket.