feat: PXE in the browser (#10353)
PXE running in a browser via a vite app. In order to achieve this,
several things had to be implemented/altered:

- Original LMDB storage was not compatible with the browser. Also its
very nature made most get operations synchronous, which is pretty weird
for a DB. A new `AztecAsyncKVStore` has been created, offering *almost*
the same interface but with async methods for everything. This store is
backed by IndexedDB, which is compatible with most modern browsers. The
LMDB backed store has been updated to also support the async interface,
so now `PxeKvDatabase` can run with either.
- In order to support testing in the browser, `@aztec/kv-store` tests
have been moved from jest to mocha
- Most imports of `fs/promises` have been replaced with `import {
promises as fs } from 'fs';` which can be polyfilled by
[vite-plugin-node-polyfills](https://www.npmjs.com/package/vite-plugin-node-polyfills).
- Several packages (such as `simulator`) have been split to allow
importing the minimal set of APIs required for PXE to do its client side
thing, avoiding unneeded (and browser-incompatible) APIs and files.
- PXE package now generates a static info file (version and name) at
compilation time, to avoid reading it from the package.json file (which
cannot be done in the browser). A bundler was considered, but this is
simpler and more flexible.
- A new flag has been added to the sandbox, so it can run without a PXE
(`--sandbox.noPXE`). This is very useful to deploy a "network in a box"
that can be used to test apps that provide their own PXE implementation.
- The vite box is very rough and mostly mimics what the `react` one
does, but shows how a hypothetical dapp could be bundled.

---------

Co-authored-by: Alex Gherghisan <alexghr@users.noreply.github.com>