chore(avm): pilcom compatibility changes (#10544)
For compatibility with upcoming design.
* Creates a columns file
* Relations now have subrelation constexpr
* Relative includes for generated files
* Now you can specify output path for generated files with `-o path`
* Full row uses macro list of columns!*

*warning: GCC didn't like this so I added
```
"CXXFLAGS": "-Wno-missing-field-initializers"
```
the reason is that I'm now defining the AvmFullRow as `FF ALL_ENTITIES;` but GCC prefers it if I do `FF col1{}; FF col2{};...` with the explicit `{}`. If I need to do that, then I need to codegen and repeat again all names.

To be honest IDK why it complains so much, if you Google this, you'll find other people running into it. This should NOT be a problem since any field not specified in aggregate initialization WILL be value-initialized. 

Now, FF's are special in that they don't initialize themselves by default, to accommodate array/other uses cases. (this is a BAD choice if you ask me, there are other ways to avoid explicit initialization if and only when you want to, especially for arrays etc).

Let's see what tests say. In our case the AvmFullRow is only used in a vector, which does somehow force full initialization of all the fields (or at least it did).