chore: Remove Proxy from json rpc client (#10554)
Using a proxy as a json rpc client means that dynamic checks for
functions in the returned object fail. For instance, `typeof
pxe.getTransactions` returns true even if the method is not part of the
pxe schema, since the proxy creates a fake function for every single
property requested. But then it fails when we try to invoke it.

Since we now have schemas, we can drop usage of the proxy and just
create an object with exactly the methods we need.