fix: Do not load pino-pretty in production bundles
Browser tests were broken since #10459 due to the static import of
pino-pretty, which had some browser unfriendly stuff. This PR removes
that import in favor of setting it up in a separate jest setup file,
only run under jest. This follows the recommendation of pino-pretty of
not importing it in production.

Note that, for end-users who want pretty logging, pino-pretty will still
kick in but as a transport, so it gets loaded in a nodejs worker thread
by pino itself, and does not pollute the static imports.

Another way of fixing this would have been adding the 'pino-pretty'
exception in the webpack fallbacks, but it felt wrong to start adding so
many deps in there.

Another option would have been to dynamically import (as opposed to
statically) pino-pretty, and use a webpackIgnore magic comment to skip
it. But it's unclear whether this would have worked with other bundlers,
and besides it would have added an async operation to our otherwise
synchronous logger setup.