feat: emulating blocks and correct (non-partial) note discovery in txe (#10356)
This is the initial approach to adding correct private note emission and
discovery in the TXe, with a tentative solution for finalized partial
notes (notes emitted in public). This enables historical proofs as well
because now we correctly use note nonces.

This is a draft PR and is looking for more feedback on
1. The general methodology
2. Partial notes implementation (incl TxEffect crafting)

Things to highlight:
1. TxHashes are hardcoded to blockNumber
3. Partial notes approach is more tenuous, we are assuming all
unencrypted logs / notes are the last note hashes, the full fix for
partials will be solved in a follow up PR if this is unsatisfactory.
4. The first nullifier is 6969 + block number. Setting this to
blockNumber only causes issues because apparently a nullifier with value
1 is already set before doing anything. See comment in code for the
issue.