feat: reduce the number unconstrained boundary hops in bloblib (#10603)
I've taken a quick look through the bloblib to see if we can make any
optimizations on the Noir code itself. This PR has a few changes from
this:

- I've removed the usage of `std::as_witness` as all the usage was
applied to the return values from unconstrained functions (and so will
be witnesses anyway).
- This shouldn't have affected performance but it also wasn't providing
us with any benefits.
- Rather than calling many unconstrained functions from a constrained
function I have created unconstrained helpers which will perform all
calculations and then return any values which are necessary for applying
constraints.
- This minimises the amount of data which gets pushed from the ACVM to
the Brillig VM and vice versa. This is important as we no longer need to
read all of arguments from calldata and cast them to the correct types
or arrange the results in memory to be read back into the main witness
map by the ACVM.
- This means that for the calculation of a single `partial_sum`, we
would have to spin up and tear down the brillig VM 17(!) times whereas
now it will be done just once.
- I've removed some cases were we'd use
`BigNum::evaluate_quadratic_expression` in order to perform simple
arithmetic and replaced them with the equivalent constrained operation.
An example for this is calculating `z_pow_d`.