chore: Remove sinon
Sinon fake timers were altering log timestamps. Instead of tweaking
the global system clock in p2p e2e tests, we introduce a date provider
that is injected as a dependency into the epoch cache (the only
component that relied on its time being in sync with L1), and we tweak
that.