chore: Remove default export for noir contracts js
To reduce long loading times when importing contract artifacts from noir
contracts js, we remove all artifacts from the default export, and
instead just expose a list of contract names available. To import a
given artifact, we now use the corresponding export, so we only need to
load that artifact and not all.