fix: always remove nullified notes (#10722)
We used to remove nullified notes as we process tagged logs, but this is
incomplete: it may happen that a recipient got no new logs but they
still need to nullify things. I imagine we never caught this because
either our tests are not comprehensive enough, of because all in all
scenarios we tried there was always a new note (e.g. a transfer for less
than the full balance resulted in a change note).

I changed things so that the note syncing process also triggers a full
search of all nullifiers for all recipients always, instead of only
doing it for those that got notes. This is perhaps not ideal long-term,
but it is a simple fix for the issue we currently have.

---------

Co-authored-by: thunkar <gregojquiros@gmail.com>