feat: Don't store every block number in block indices DB (#10658)
The block indices DB is what enables clients to query the block number
for a given index. Where there were blocks of zero size (so multiple
block numbers against an index) it was storing every block number
against that index. This could result in an unbounded values size in the
DB. We now just store the block range in this scenario.