fix: Do not abort blockstream loop during sync (#10884)
The L2 blockstream is expected to download blocks continuously until
reaching the source's block number. However, to exit from that loop when
the service is stopped, we continuously check for the isRunning flag and
abort if it is not set.

This means that in a sync, which happens outside the running loop, we
end up cutting that loop short. This provides a quick and dirty fix for
that.