feat: Reduce bundle sizes, remove polyfills (#10877)
Cleans up debug metadata for contract artifacts in `protocol-contracts`
and `noir-protocol-circuit-types`, adding an optional build task to keep
them. This heavily impacts the amount of data that we import in other
parts of the code.

Also cleans up some testing imports that forced us to use weird vite
plugins to conditionally delete blocks of code.

Starts the very much needed code splitting in `circuits.js` isolating
everything that has to do with blobs and part of the rollup-only types.

Finally, this allows us to create a minimal `vite.config.ts` file that
only polyfills `path`, `buffer` and `process` even with PXE proving in
the browser! Bundle sizes of this minimal app have gone from 100MB raw,
45 compressed -> 50MB raw, 30 compressed.