fix: Conditionally deploy deterministic deployment proxy
The boot-node init script sent the deterministic deployment proxy [deployment tx](https://github.com/Arachnid/deterministic-deployment-proxy?tab=readme-ov-file#deployment-transaction) always, regardless of the code being present or not. If the tx had already been sent, this caused the command to fail with a `nonce too low`. This commit adds a check so that it only sends the tx if the proxy has not being deployed yet.