chore: Faster getLeafOrLowLeafInfo in AVM ephemeral trees
The implementation of `_getLeafOrLowLeafInfo` would search for the low
leaf info in both the ephemeral indexed updates and on the actual db. If
it didn't get an exact hit for the leaf, it would start from the hit on
the indexed updates, and then linearly walk through the underlying
nullifier tree until it got to the low leaf.

This PR changes it so queries low leaf from both indexed and external,
and picks the closest one.