chore: Cache files downloaded from cache
To avoid a roundtrip to S3 to re-download every artifact on
cache_download, we store a local file with the file prefix that points
to the latest downloaded file version (ie hash), so we don't re-donwload
it we already have it.

This required changing the arity of the cache_download and cache_upload
scripts, so they accept the key and version for each artifact
separately.

In addition, to avoid issues with the nargo `show-program-hash` being
indeterministic, we also store "symlink" files in the cache for each
circuit and vk. These are text files named after the entire hash of the
circuits folder, that contain the nargo-computed hash of the circuit.
This acts as another key for the same artifact.