chore: Add afterEach hook to log when a test finishes
We have a beforeEach hook in e2e tests that lets us know in the logs
when a test starts, but we don't have an afterEach that points to when
it finishes. Note that we cannot use the start of the next test as a
pointer, since other setup hooks get in the way.