chore: Bootstrap cache download ignores PARALLELISM
We had introduced the PARALLELISM env var for bootstrap in #10909 so we
could build on non-mainframe machines. However, setting a low
parallelism means that requests to the cache when building noir circuits
happen serially, so latency now becomes a problem.

This PR is an attempt to set up a pipeline of parallel operations, where
we first run all cache downloads with very high parallelism, output what
needs to be built, and run that with a lower parallelism to avoid
killing the host.