chore: Runs bb and nargo behind sem in bootstrap
This PR removes the PARALLELISM for building each contract and circuit
that had been introduced in #10909. Instead, it adds a `sem` around the
resource-hungry sections of code, using the PARALLELISM default. This
allows for more parallelism of the entire script, while slowing down
only on the parts where it's really needed.

Alternative to #11030.