fix: Check class registration nullifier in node before returning class (#11074)
Since https://github.com/AztecProtocol/aztec-packages/issues/10000 we
are not broadcasting public bytecode. To work around this in some
networks, we're manually adding some common contracts (Token) to
archivers on startup.

Since an Aztec node returns that a class is publicly registered as long
as it's on its archiver, clients were not sending txs to register the
Token class.

However, we will soon check for registration nullifiers. But since
clients think the Token class is registered, they never send the
registration tx. So this will cause all interactions with the Token to
break.

This attempts to fix this by only telling clients that a class is
registered if the nullifier is present.