feat: Validate block proposal txs iteratively (#10921)
Instead of loading all txs from the p2p pool and validating them all, we
now get an iterator to the p2p pool and iteratively run through them and
validate them as we go. This ensures we only load and validate strictly
the txs we need. This also makes it easy to enforce new block
constraints such as gas limits, which we add as two new env vars.

As part of this PR, we also change the interface of validators. Since
there is no point anymore in validating txs in bulk, we drop the
`validateTxs` method in favor of just `validateTx`. And since we're at
it, we enrich `validateTx` to return `valid/invalid/skip` and to include
the failure reason.

Fixes #10869