chore!: Attestation collection times out based on sequencer timetable
Removes the env var `VALIDATOR_ATTESTATIONS_WAIT_TIMEOUT_MS` and
replaces it with the deadline for publishing the block to L1. This
ensures the sequencer does not get stuck waiting for attestations where
the validators have given up due to a reexecution timeout.

If timetable is not enforced, this value defaults to one aztec slot.