fix: verify start state of a block (#11290)
Check the start state of the first tx in a block, in block root rollup
circuit, to make sure that it matches the state in the previous block
header, whose hash is the last leaf in the last archive. Otherwise, it
could be whatever and that allows adding fake data by changing the tree
roots in the start state.