chore: Try fix e2e block building flake (#11359)
The test was failing [in some
runs](https://github.com/AztecProtocol/aztec-packages/actions/runs/12869132447/job/35878509644#step:3:4657)
because apparently the 32 txs take too long to go in (they are processed
1 or 2 per block only) and a reorg kicks in before that.

This PR reduces the total number of txs so we don't reach that point.
Note that if we speed up tx processing significantly, then all txs could
end up processed within one second and end up on the same block. If that
happens, we'll need to increase the number of txs again.