feat: check contract classes from revertible/non-revertible logs are handled correctly (#11396)
### Re: a discussion on supporting non-revertible contract class logs
(no issue num)

This PR ensures we only remove contract classes (and instances) from
revertible side effects when we have a revert. The changes only affect
the public `classCache` and `instanceCache`, not the underlying
`dataSource` (which is used in `getContractClass` as a backup).
Even though I didn't touch any other data sources, the new test I added
implies this is working fine i.e. revertible contract classes are
removed, and non-revertible are kept.

The test in noir is pretty ugly and the result of many attempts to build
a test that:

- emits a contract class log successfully
- emits the log such that it can be set as revertible or non-revertible
- also makes a public call which reverts

Suggestions on how to improve it are very welcome!!