feat!: leonidas library (#11596)
Fixes #11521, slaying the king. Pushes a bunch of the functions that
lived in the "rollup" our from the core, such that there now is a
`RollupCore` with the logic that we need to run the rollup and `Rollup`
extending it with a bunch of helpers and such.

TBH, the `RollupCore` is still pretty disgusting to look at, but some of
this should be helped as we split the state transition function checks
further from the sequencer selection. Will alos allow us to get rid of
some of the existing bundling libraries and improve clarity.


Makes a tiny update in the forwarder and one of the tests for it.