fix: contracts with no private / public functions should not fail to generate a contract artifact (#11744)
If there were no public / private functions, the abi structs for
functions is undefined. In the previous case, with any private / public
function, unconstrained functions would just not exist in the
abistructs, and returnTypes would stay an empty array, but if there were
no private / public functions, abiStructs would be undefined and we
would throw an error. This fixes that.