feat: indexed protocol contracts tree (#11897)
Closes #11843

This PR converts the protocol contract address tree to an indexed tree.
This is so:
- If a user is calling a protocol contract address, its calculated
address must be in the tree
- If a user is NOT calling a protocol contract address, its calculated
address must NOT be in the tree

This required more changes than expected because the old code had this
statement:
```rust
    let computed_protocol_contract_tree_root = if (MAX_PROTOCOL_CONTRACTS as Field).lt(
        protocol_contract_index,
    ) {
        0
    } else {
        root_from_sibling_path(
            computed_address.to_field(),
            protocol_contract_index,
            private_call_data.protocol_contract_sibling_path,
        )
    };
    assert(
        computed_address.eq(contract_address)
            | computed_protocol_contract_tree_root.eq(protocol_contract_tree_root),
        "computed contract address does not match expected one",
    );
```
For many tests in and out of nr, we input `protocol_contract_tree_root =
0`. That meant that even if we weren't calling a protocol contract
address, an incorrect `contract_address` would pass the assert
statement. The current code does not allow this, so I had to update some
old fixtures and provide a real root for all calls (not just protocol
contract ones).