chore: @aztec/stdlib pt1 -> cleanup circuits js (#12039)
This initial PR focuses on:

- Moving generic stuff from `circuits.js` to `foundation`: merkle tree
calculator is general enough to be on foundation, same with crypto
functions.
- Extracting references to blobs from `circuits.js` so they're not
accidentally imported into browser bundles again. Still work to do here,
this will only be completely realized once the "catchall" export is
removed.
- Extracting `constants` to its own package to deal with circular
references -> `blob-lib` and `circuits.js` need them, this opens the
door to cleaner and more advanced anti "I forgot to update constants"
measures.
- Prepare `foundation` and `circuits.js` so aztec-specific code is moved
from the former to the latter (will be done in a new PR, specifically
targetting aztec-address, abi already moved).
- Preparing `circuits.js` for a reorg, specifically targeting the
removal of the "catchall" export and moving to granular approach. This
will end with the renaming of `circuits.js` to `stdlib`.
- Preparing for the removal of a lot of reexports. They will be kept to
a minimum, with the exception of outwards-facing APIs such as `aztec.js`