chore: Do not allow changes to package json during CI yarn install
Running `yarn install` will format the package.json for all packages in
the workspace. This will cause `cache_content_hash` to fail on CI due to
changed files, which causes the build to terminate abruptly (see
[here](https://github.com/AztecProtocol/aztec-packages/actions/runs/13422411604/job/37497960586?pr=12096#step:7:109) for an example).

This PR adds all package.json files to yarn's `immutablePatterns`, so if
it detects that the install process changed them it, it will fail with
`The checksum for **/package.json has been modified by this install,
which is explicitly forbidden.` and a nice exit code, which should be
visible on the CI run. This only applies to CI runs.