fix: Fix epoch monitoring and related e2e tests (#12096)
Fixes the prover-node's epoch monitor so it accounts for the previous
epoch to be proven. The updated logic is that an epoch is ready to prove
only if it is completed and it contains the first unproven block (which
gets updated over a reorg).

Also fixes #11840 

Builds on top of @PhilWindle's #12014

---------

Co-authored-by: PhilWindle <philip.windle@gmail.com>