chore: use cheatcodes instead of calling directly (#12047)
Fixes #12001. 

Biggest changes:
- Removes the `assumeProvenThrough`, no more calls to the contract to
update that one
- Use anvil cheatcodes to directly overwrite the storage and update the
proven block number
- Adding task on the anvil test watcher, to mark blocks as proven
  - Can be toggled by updating `isMarkingAsProven`
- For example see the fee tests or the e2e_block building test that is
trying to re-org

The watcher chances could feel a little clunky. But as we are in most
cases not testing the proving, this felt like the best way to make a
sane default ("automining") and then require an extra step for the case
of using a prover node or no proving at all.