fix: Archiver can sync from scratch with pruneable blocks
Archiver was unable to sync successfully when there were blocks awaiting
to be pruned. This was caused by:

- An error when fetching blocks, that caused them to be returned and
processed out-of-order. This caused the "last processed L1 mark" to be
incorrect, leading to the same L1 blocks to be downloaded more than
once in subsequent archiver iterations. This is now fixed.

- An overzealous block store, that blew up when requesting it to unwind
  a block it didn't have. It now just warns.

- The archiver downloading blocks that are unwound immediately after,
  since the check for pruning is done after blocks download. This is not
fixed in this PR, since naively adding a check for pruning _before_ blocks are
downloaded would mean significant more calls to `Rollup.status`, since
it would get called every time and not just when there are new blocks.
Not fixing this means that, in an initial sync, some additional blocks
(up to 2 epochs worth) get downloaded. Not too much of an issue.

Fixes #12206