chore: Add logger optional arguments to bb entrypoints
Adds a logger optional argument to bb classes and functions so that the
default `debug` logger can be replaced with something else defined by
the caller. We can then tweak this from aztec-packages to use the pino
logger.

Also cleans up logs in the CLI a bit.

Also ensures the `wasmPath` is actually used for loading the wasm when
it is specified.