chore: Lazy loading artifacts everywhere (#12285)
This PR adds the *option* to lazily load JSON artifacts in all the
packages that were missing that functionality, further improving bundle
sizes for the browser and allowing us to finally put limits on our
example apps bundle sizes:

* `noir-protocol-circuits-types/vks`: now allow vk lazy imports via the
`LazyArtifactProvider` in the `/client` export, while maintaining the
bundled version for the server under `/server/vks`
* `accounts`: Now all exports provide a lazy version, e.g:
`@aztec/accounts/schnorr/lazy`. This has proven to be complicated due to
the testing import leaking into the browser (where type assertions are
not widely supported). Now there's also `/testing/lazy`. This testing
package is needed in the browser for the prefunded accounts.
* `protocol-contracts`: Now with a lazy version and exporting
`ProtocolContractProviders` so that PXE can be configured with bundled
and lazy versions.


Besides the type assertion issue, we absolutely want to keep bundled and
lazy versions because some environments don't play well with `await
import` (namely service workers in firefox)

---------

Co-authored-by: esau <152162806+sklppy88@users.noreply.github.com>