feat: extend functionality of SmallSubgroupIPA  (#12244)
To fix the issue with translation evaluations when the masking is
enabled in `ECCVM`, we need to prove the batched evaluation of the
masking term, where by masking term we mean a small table (5 columns, 4
rows) added at the end of the Transcript polynomials.

Fortunately, we could use the SmallSubgroupIPA protocol to do this
efficiently and faciliatate the integration of this fix into Goblin.

This is the first batch of changes: 
* Introduction of a class `TranslationData` designed to process the
masked wires and commit to the masking term
* Extension of `SmallSubgroupIPA` functionality - now it could accept
Translation data and produce a proof of the batch opening of the masking
term
* Added corresponding tests to `SmallSubgroupIPATests`
* Moved implementations to .cpp