feat(config): add fallbacks (#12593)
## Overview

Again like https://github.com/AztecProtocol/aztec-packages/pull/12589
this is added after a chat with
charlie where we want to remove the need for a complex script. 

https://github.com/AztecProtocol/aztec-packages/pull/12566 asks the user
for a private key as it needs
to set two env vars to the same key. 

This pr introduces the notion of a fallback, such that a config variable
will also check a list of fallbacks before
using the default value.

In this case, if `VALIDATOR_PRIVATE_KEY` is set, then
`SEQ_PUBLISHER_PRIVATE_KEY`` does not need to be set, removing the
need for it in the script.