feat: slim wallet (#12803)
Reduces the surface of our `Wallet` API, focusing on removing the
extension of `PXE`. This went downhill fast, but I think that ended up
better than it was:

``` typescript
/**
 * The wallet interface.
 */
export type Wallet = AccountInterface &
  Pick<
    PXE,
    // Simulation
    | 'simulateTx'
    | 'simulateUnconstrained'
    | 'profileTx'
    // Sending
    | 'sendTx'
    // Contract management (will probably be collapsed in the future to avoid instance and class versions)
    | 'getContractClassMetadata'
    | 'getContractMetadata'
    | 'registerContract'
    | 'registerContractClass'
    // Likely to be removed
    | 'proveTx'
    // Will probably be collapsed 
    | 'getNodeInfo'
    | 'getPXEInfo'
    // Fee info
    | 'getCurrentBaseFees'
    // Unclear 
    | 'updateContract'
    // Sender management
    | 'registerSender'
    | 'getSenders'
    | 'removeSender'
    // Tx status
    | 'getTxReceipt'
    // Events. Kept bc events are going to be reworked and removing them now increases the already huge diff
    | 'getPrivateEvents'
    | 'getPublicEvents'
  > & {
    createAuthWit(intent: IntentInnerHash | IntentAction): Promise<AuthWitness>;
  };
```

- Authwits are not longer stored in PXE. It really didn't make sense:
either they were used and "consumed" (or at least they should in most
use cases), or the txs failed because they were invalid and as such
didn't make sense to keep them around. Removing them from PXE's DB and
forcing them to be provided externally cleans up the wallet's interface
considerably, and delegates authwit management on the wallet software
(which I think it's good). Closes:
https://github.com/AztecProtocol/aztec-packages/issues/12394
- The above meant supporting transient authwits in unconstrained
simulation, which has been done.
- Removes a bunch of debug utilities from wallet (namely `debug` option
from `.send` and `proven: true`) In the real world it doesn't make sense
to wait for several epochs so txs are proven (a utility is kept for e2e
testing) and all the info that `debug` retrieved can be easily obtained
via a single query to the node. This also cleans up the interface
nicely.
- Cleans up the `@aztec/entrypoints` package, attempting to reduce the
interdependency with `aztec.js`
- Starts addressing the insanity that is `ContractFunctionInteraction`.
There's a new object called `ExecutionPayload` that represents the
"minimal unit of execution" in Aztec. It consists of contract calls,
authwits and capsules, and every single higher abstraction object in our
codebase generates one. After that, they can easily be composed to
create complex interactions, such as account entrypoints or batched
calls. It builds up on the system @LeilaWang designed, but starts a
little bit earlier in the chain and I think ends up looking cleaner.
- As such, a lot of the complexity of our
DeployMethod/DeployAccountMethod classes have been delegated on these
objects and the factories that build them.

---------

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>